requirement.txt
amqp==2.2.2
billiard==3.5.0.3
boto3==1.4.7
botocore==1.7.48
celery==4.1.0
Django==2.0.4
django-filter==1.1.0
django-redis==4.9.0
djangorestframework==3.8.1
docutils==0.14
jmespath==0.9.3
kombu==4.1.0
Markdown==2.6.11
pycurl==7.43.0
python-dateutil==2.7.2
pytz==2018.3
redis==2.10.6
s3transfer==0.1.13
six==1.11.0
vine==1.1.4
